requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
passlib==1.7.4
psycopg2-binary==2.9.9
python-dotenv==1.0.0
email-validator==2.1.0.post1
bcrypt==4.1.2
python-multipart==0.0.16
python-jose==3.3.0
python-email-validator==2.0.0

